requirements.txt
numpy==1.26.2
mergedeep==1.3.4
mpi4py==3.1.3
matplotlib==3.5.1
PyYAML==6.0
